kill rails default index.html; leaderboard_line_items ranking scope; switch leaderboard_line_items metric datatypes to int; introduce ClubContestController and associated views/styling

+++ db/migrate/20120512184946_create_leaderboard_line_items.rb
@@ -2,9 +2,9 @@ class CreateLeaderboardLineItems < ActiveRecord::Migration
   def change
     create_table :leaderboard_line_items do |t|
       t.references :leaderboard
-      t.float :average_elevation
+      t.integer :average_elevation
       t.integer :number_of_rides
-      t.float :elevation_gain
+      t.integer :elevation_gain
       t.string :athlete_name, :limit => 60
 
       t.timestamps
